requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
pydantic
requests